requirements.txt
aioice==0.9.0
aiortc==1.6.0
altair==5.2.0
attrs==23.2.0
av==11.0.0
backports.zoneinfo==0.2.1;python_version<"3.10"
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.0
dnspython==2.5.0
gitdb==4.0.11
GitPython==3.1.41
google-crc32c==1.5.0
idna==3.6
ifaddr==0.2.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
numpy==1.24.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pkgutil-resolve-name==1.3.10
protobuf==4.25.2
pyarrow==15.0.0
pycparser==2.21
pydeck==0.8.1b0
pyee==11.1.0
pygments==2.17.2
pylibsrtp==0.9.1
pyOpenSSL==24.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
six==1.16.0
smmap==5.0.1
streamlit==1.30.0
streamlit-webrtc==0.47.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing-extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
